% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.fun.R
\name{med.fun}
\alias{med.fun}
\title{Internal med.fun function}
\usage{
med.fun(y.data, index, m.data)
}
\arguments{
\item{y.data}{not info needed}

\item{index}{not info needed}

\item{m.data}{not info needed}
}
\value{
not info needed
}
\description{
This is an internal function copied from the mediation package
(mediation:::med.fun)
}
\details{
Not required
}
\keyword{internal}
