% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neImputeDS.R
\name{neImputeDS}
\alias{neImputeDS}
\title{Expand the dataset and impute nested counterfactual outcomes}
\usage{
neImputeDS(object)
}
\arguments{
\item{object}{a string character, the name of an object used to select a method.}
}
\value{
a data frame of class c("data.frame", "expData", "impData") is assigned at
the server-side of each study.
}
\description{
This function is similar to R function \code{neImpute} from the 
\code{mmedflex} package.
}
\details{
The function 'neImpute' both expands the data along hypothetical exposure
values and imputes nested counterfactual outcomes.
}
\author{
Demetris Avraam, for DataSHIELD Development Team
}
