% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neLhtDS.R
\name{neLhtDS}
\alias{neLhtDS}
\title{Linear hypotheses for natural effect models}
\usage{
neLhtDS(model.name)
}
\arguments{
\item{model.name}{a fitted natural effect model object. This is the object saved
on the server-side by the \code{ds.neModel} function.}
}
\value{
a summary table of the object of class c("neLht", "glht") (see glht).
}
\description{
This function is similar to R function \code{neLht} from the 
\code{medflex} package.
}
\details{
The function \code{neLhtDS} allows to calculate linear combinations of 
natural effect model parameter estimates.
}
\author{
Demetris Avraam, for DataSHIELD Development Team
}
