% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synDS.R
\name{synDS}
\alias{synDS}
\title{Generating synthetic datasets}
\usage{
synDS(data, method, m = 1, k = NA, proper = FALSE, seed = NA)
}
\arguments{
\item{data}{a dataframe containing the variables to be included in the synthetic dataset}

\item{method}{a single string or a vector of strings of length ncol(data)
specifying the synthesising method to be used for each variable in the data.
Order of variables is exactly the same as in data.}

\item{m}{number of synthetic copies of the original (observed) data to be generated. The default is m = 1.}

\item{k}{number of rows to generate}

\item{proper}{a logical value with default set to FALSE. If TRUE proper synthesis is conducted.}

\item{seed}{an integer to be used as an argument for the set.seed(). If no integer is provided, one will be generated}
}
\value{
synthetic datasets
}
\description{
This function is similar to R function \code{syn} from the 
\code{synthpop} package.
}
\details{
The function 'syn' is used to generate synthetic versions of a dataset.
}
\author{
Tom Bishop
}
