% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synDS2.R
\name{synDS2}
\alias{synDS2}
\title{Generating synthetic datasets}
\usage{
synDS2(data, ...)
}
\arguments{
\item{data}{a dataframe containing the variables to be included in the synthetic dataset}

\item{...}{the arguments passed to \code{syn}}
}
\value{
synthetic datasets
}
\description{
This function is similar to R function \code{syn} from the 
\code{synthpop} package.
}
\details{
The function 'syn' is used to generate synthetic versions of a dataset.
The following disclosure protections are enforced
}
\author{
Tom Bishop
}
